+++ exploration/metadata_documents.py
@@ -2,7 +2,7 @@ import random
 import itertools
 from collections import Counter
 
-from utility import find_id_from_value
+from exploration_utility import find_id_from_value
 import preprocess.preprocessing as pre
